Update TeleOp_H2OLooBots.java
We do not have a limelight on our bot. Until we begin to work on automatic angle adjustment, I removed the basic limelight code so we don't have hardware map issues when developing the base teleop file. It was the most simply code and didnt actually do anything but start the limelight, so we can easily add it back later effortlessly. - kyeler